When parsing, reset tag type to null between tags.
This eliminates an issue where because of missing braces in a typed tag,
the type of the previously processed tag could be used for the current
tag. See #25.

This was causing a crash in a case where an @extend was missing braces,
and the previous tag was a constructor argument with the same type as
the constructor itself, so that the class was set up as its own
superclass... eventually leading to a stack overflow. This corner case
has prevented my effective use of this plugin at work.

This is by far the worst I have ever been bitten by JS variable scoping.