Load core Tern modules on initialize().
Rather than blindly trying to require() the core Tern modules and hoping
we load the same modules that are loading the plugin, we wait for Tern
to call initialize() on the main tern-closure module, telling us where
to load the Tern core from. We track references to the core modules in a
simple module, lib/core.js.

This fixes potential installation issues, particularly for users
following the Tern setup flow for Emacs.